ml_engine/requirements_serving.txt

fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
torch==2.1.2 --index-url https://download.pytorch.org/whl/cpu
transformers==4.37.0
easyocr==1.7.1
pillow==10.2.0
numpy==1.26.3
